Update patch set 2
Patch Set 2:

Ah ok, got it. Usually npm is bundeled with node.js and i assumed that npm is also bundeled in the sipwise nodejs debian backport. But in jessie there are two separate packages. So, either we install npm as debian package in addition to the nodejs backport or we install nodejs+npm via the bash script. In my opinion a clean solution would be to support a npm backport that matches the corresponding nodejs backport. But i think the solution with min effort is to use the bash script.

Patch-set: 2
Reviewer: Hans-Peter Herzog <1000026@85315e57-5f03-49c3-83f8-201ae2313a75>
Label: Code-Review=0