TT#113853 Preparations for release mr8.5.4, missing "mocha" binary
Similar to what happened with janus-admin, the job for unit test fails:

  23:57:55 > janus-videoroom-client@4.1.7 test /tmp/code
  23:57:55 > mocha -b --exit -R spec --full-trace test/*-spec.js
  23:57:55
  23:57:55 sh: 1: mocha: not found
  23:57:55 npm ERR! Test failed.  See above for more details.

This dependency seems to be present in packages*.json package, however
it doesn't work.

So adding mocha as a pakcage to install, to try to get around the
problem and be able to build the package for the mr8.5.4 release.

Change-Id: Id3464e0e9f8089b43048daa1c7941b06ad706bab